chore: Move functions currently in types files to types.ts
