+++ indexer/tests/header.test.ts
@@ -6,11 +6,8 @@ import {
   PrefixedHexString,
 } from "../src/deps.ts";
 import { assertEquals } from "https://deno.land/std@0.213.0/assert/assert_equals.ts";
-import {
-  DEFAULT_BLOCK_GAS_LIMIT,
-  JsonRpcBlock,
-  toEthHeader,
-} from "../src/types/header.ts";
+import { DEFAULT_BLOCK_GAS_LIMIT, toEthHeader } from "../src/types/header.ts";
+import { JsonRpcBlock } from "../src/types/types.ts";
 import { padString } from "../src/utils/hex.ts";
 import sinon from "npm:sinon";
 import { KAKAROT } from "../provider.ts";
